Updating based on e8dcdbfdc

+++ resource-manager/insights/2024-02-01/metrics/model_response.go
@@ -4,7 +4,7 @@ package metrics
 // Licensed under the MIT License. See NOTICE.txt in the project root for license information.
 
 type Response struct {
-	Cost           *float64 `json:"cost,omitempty"`
+	Cost           *int64   `json:"cost,omitempty"`
 	Interval       *string  `json:"interval,omitempty"`
 	Namespace      *string  `json:"namespace,omitempty"`
 	Resourceregion *string  `json:"resourceregion,omitempty"`
